Move observers from utils to queries
