+++ app/components/post_draft/send_handler/index.ts
@@ -10,9 +10,9 @@ import {General, Permissions} from '@constants';
 import {MAX_MESSAGE_LENGTH_FALLBACK} from '@constants/post_draft';
 import {observeChannel, observeCurrentChannel} from '@queries/servers/channel';
 import {queryAllCustomEmojis} from '@queries/servers/custom_emoji';
+import {observePermissionForChannel} from '@queries/servers/role';
 import {observeConfig, observeCurrentUserId} from '@queries/servers/system';
 import {observeUser} from '@queries/servers/user';
-import {observePermissionForChannel} from '@utils/role';
 
 import SendHandler from './send_handler';
